services/auth_service/requirements.txt
grpcio==1.78.1
grpcio_tools==1.78.1
passlib==1.7.4
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
PyJWT==2.11.0
redis==7.2.0